requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0

# Optional: for future enhancements
# selenium>=4.15.0  # For browser automation if needed
